verilog-blink: fix when building under cygwin
When building under cygwin, we are building under windows, however the
SHELL is not cmd.exe.  As a result, our replacement commands such as
`copy` don't work.

Add an additional check to see if `sh.exe` is usable, and if it is don't
replace the functions.

This fixes the build when using a posix-ish environment under Windows,
such as cygwin or msys2.

This fixes issue #69.

Signed-off-by: Sean Cross <sean@xobs.io>