renode-bridge: use wishbone-tool as the main bridge
litex is very volatile, and tends to reorganize its submodules.  This
can lead to confusion when doing a `git pull`.  For example:

https://github.com/im-tomu/fomu-workshop/issues/180

Additionally, instructions for running `litex_server.py` are different
depending on the platform.

This unifies the platforms by replacing `litex_server.py` with
`wishbone-tool`, which users will already have installed anyway.

Signed-off-by: Sean Cross <sean@xobs.io>